cleanupTestFile should not fail with NPE if testFile is null because of another reason in unit test

+++ delis-web/src/test/java/dk/erst/delis/task/document/TestDocumentUtil.java
@@ -29,6 +29,9 @@ public class TestDocumentUtil {
 	}
 
 	public static void cleanupTestFile(Path testFile) {
+		if (testFile == null) {
+			return;
+		}
 		if (testFile.toFile().exists()) {
 			if (testFile.toFile().delete()) {
 				log.info("Deleted test file " + testFile);
